+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KF08B1ZFD6D8CZWAX1QNNQ3K"
+        "01KF0T8PK1SYVSSQ419T4ZYK8D"
     )
     r"""The unique identifier of the data source"""