+++ packages/orq-rc/src/orq_ai_sdk/models/updateidentityop.py
@@ -194,7 +194,7 @@ class UpdateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-15T07:18:41.663Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-15T12:31:58.759Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")